test removeMetadata

+++ syncer/syncer.go
@@ -426,7 +426,6 @@ func (s *Syncer) initShardingGroups() error {
 
 // IsFreshTask implements Unit.IsFreshTask
 func (s *Syncer) IsFreshTask(ctx context.Context) (bool, error) {
-	// TODO(lance6716): add check for downstream table select count not zero, and load that checkpoint
 	globalPoint := s.checkpoint.GlobalPoint()
 	tablePoint := s.checkpoint.TablePoint()
 	return binlog.CompareLocation(globalPoint, binlog.NewLocation(s.cfg.Flavor), s.cfg.EnableGTID) <= 0 && len(tablePoint) == 0, nil
@@ -1092,18 +1091,14 @@ func (s *Syncer) Run(ctx context.Context) (err error) {
 		}
 
 		// for fresh and all-mode task, flush checkpoint so we could delete metadata file
-		if s.cfg.Mode == config.ModeAll {
-			if err = s.flushCheckPoints(); err != nil {
-				s.tctx.L().Warn("fail to flush checkpoints when starting task", zap.Error(err))
-			} else if s.cfg.CleanDumpFile {
-				s.tctx.L().Info("try to remove loaded files")
-				metadataFile := path.Join(s.cfg.Dir, "metadata")
-				if err = os.Remove(metadataFile); err != nil {
-					s.tctx.L().Warn("error when remove loaded dump file", zap.String("data file", metadataFile), zap.Error(err))
-				}
-				if err = os.Remove(s.cfg.Dir); err != nil {
-					s.tctx.L().Warn("error when remove loaded dump folder", zap.String("data folder", s.cfg.Dir), zap.Error(err))
-				}
+		if s.cfg.Mode == config.ModeAll && s.cfg.CleanDumpFile {
+			s.tctx.L().Info("try to remove loaded files")
+			metadataFile := path.Join(s.cfg.Dir, "metadata")
+			if err = os.Remove(metadataFile); err != nil {
+				s.tctx.L().Warn("error when remove loaded dump file", zap.String("data file", metadataFile), zap.Error(err))
+			}
+			if err = os.Remove(s.cfg.Dir); err != nil {
+				s.tctx.L().Warn("error when remove loaded dump folder", zap.String("data folder", s.cfg.Dir), zap.Error(err))
 			}
 		}
 	}
